+++ test/src/ProcessManagerTest.php
@@ -107,7 +107,7 @@ class ProcessManagerTest extends TestCase
         $this->assertSame($manager, $result);
         $this->assertSame($callback, $this->extractProperty($manager, 'processStartCallback'));
     }
-    
+
     /**
      * Tests the setProcessFinishCallback method.
      * @covers ::setProcessFinishCallback
@@ -199,17 +199,32 @@ class ProcessManagerTest extends TestCase
     }
 
     /**
-     * Tests the executeNextPendingProcess method.
-     * @covers ::executeNextPendingProcess
-     * @throws ReflectionException
+     * Provides the data for the executeNextPendingProcess test.
+     * @return array
      */
-    public function testExecuteNextPendingProcess(): void
+    public function provideExecuteNextPendingProcess(): array
+    {
+        return [
+            [42, true, false],
+            [null, false, true],
+        ];
+    }
+
+    /**
+     * Tests the executeNextPendingProcess method.
+     * @param int|null $pid
+     * @param bool $expectRunningProcess
+     * @param bool $expectCheck
+     * @throws ReflectionException
+     * @covers ::executeNextPendingProcess
+     * @dataProvider provideExecuteNextPendingProcess
+     */
+    public function testExecuteNextPendingProcess(?int $pid, bool $expectRunningProcess, bool $expectCheck): void
     {
         $processStartDelay = 1337;
         $callback = 'strval';
         $processStartCallback = 'intval';
         $env = ['foo' => 'bar'];
-        $pid = 42;
 
         /* @var Process|MockObject $process */
         $process = $this->getMockBuilder(Process::class)
@@ -231,17 +246,12 @@ class ProcessManagerTest extends TestCase
             ['abc', 'def'],
         ];
 
-        $runningProcesses = [
-            1337 => 'ghi',
-        ];
-        $expectedRunningProcesses = [
-            1337 => 'ghi',
-            42 => $process,
-        ];
+        $runningProcesses = [1337 => 'ghi'];
+        $expectedRunningProcesses = $expectRunningProcess ? [1337 => 'ghi', 42 => $process] : ['1337' => 'ghi'];
 
         /* @var ProcessManager|MockObject $manager */
         $manager = $this->getMockBuilder(ProcessManager::class)
-                        ->setMethods(['canExecuteNextPendingRequest', 'sleep', 'invokeCallback'])
+                        ->setMethods(['canExecuteNextPendingRequest', 'sleep', 'invokeCallback', 'checkRunningProcess'])
                         ->setConstructorArgs([0, 0, $processStartDelay])
                         ->getMock();
         $manager->expects($this->once())
@@ -250,6 +260,10 @@ class ProcessManagerTest extends TestCase
         $manager->expects($this->once())
                 ->method('sleep')
                 ->with($processStartDelay);
+        $manager->expects($expectCheck ? $this->once() : $this->never())
+                ->method('checkRunningProcess')
+                ->with($pid, $process);
+
 
         $manager->expects($this->once())
                 ->method('invokeCallback')
@@ -337,23 +351,29 @@ class ProcessManagerTest extends TestCase
     public function provideCheckRunningProcess(): array
     {
         return [
-            [true, false],
-            [false, true],
-            [false, true],
+            [42, true, false, false],
+            [42, false, true, true],
+            [42, false, true, true],
+            [null, false, true, false],
         ];
     }
 
     /**
      * Tests the checkRunningProcess method.
+     * @param int|null $pid
      * @param bool $resultIsRunning
      * @param bool $expectFinish
+     * @param bool $expectUnset
      * @throws ReflectionException
      * @covers ::checkRunningProcess
      * @dataProvider provideCheckRunningProcess
      */
-    public function testCheckRunningProcess(bool $resultIsRunning, bool $expectFinish): void
-    {
-        $pid = 42;
+    public function testCheckRunningProcess(
+        ?int $pid,
+        bool $resultIsRunning,
+        bool $expectFinish,
+        bool $expectUnset
+    ): void {
         /* @var Process|MockObject $process */
         $process = $this->getMockBuilder(Process::class)
                         ->setMethods(['checkTimeout', 'isRunning'])
@@ -368,7 +388,7 @@ class ProcessManagerTest extends TestCase
         /* @var Process $process2 */
         $process2 = $this->createMock(Process::class);
         $runningProcesses = [42 => $process, 1337 => $process2];
-        $expectedRunningProcesses = $expectFinish ? [1337 => $process2] : [42 => $process, 1337 => $process2];
+        $expectedRunningProcesses = $expectUnset ? [1337 => $process2] : [42 => $process, 1337 => $process2];
 
         /* @var ProcessManager|MockObject $manager */
         $manager = $this->getMockBuilder(ProcessManager::class)