+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Add
 
 - Option to delay the start of processes to space them out.
+- `ProcessManagerInterface`.
 
 ## 1.0.0 - 2018-09-10