+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Option to delay the start of processes to space them out.
 - `ProcessManagerInterface`.
 
+### Fixed
+
+- Failing to always retrieve PID on immediately completing processes.
+
 ## 1.0.0 - 2018-09-10
 
 - Initial version of the process manager.